feat: Names 객체 생성 및 분리
